+++ NEWS.md
@@ -1,6 +1,8 @@
 # phenoptrReports 0.1.1.9000
 
 New features:
+- Mean expression sheet includes tissue category "All" with mean expression
+  across all requested categories.
 - Chart report includes count and density charts with truncated Y-axis 
   if any count or density > 2000